Add csrf token for delete

+++ src/Controller/PinsController.php
@@ -16,53 +16,51 @@ class PinsController extends AbstractController
     /**
      * @Route("/", name="app_home")
      */
-    public function index(PinRepository $pinRepository): Response 
+    public function index(PinRepository $pinRepository): Response
     {
-        $pins = $pinRepository->findBy([], ['createdAt' =>'DESC' ]);
-       
+        $pins = $pinRepository->findBy([], ['createdAt' => 'DESC']);
+
         return $this->render('pins/index.html.twig', compact('pins'));
     }
 
-     /** 
-    *@Route("/create", name="app_pins_create", methods={"GET", "POST"})
-    */
+    /** 
+     *@Route("/create", name="app_pins_create", methods={"GET", "POST"})
+     */
     public function create(Request $request, EntityManagerInterface $em): Response
     {
 
         $pin = new Pin;
-       
+
         $form = $this->createForm(PinType::class, $pin);
-        
+
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid())
-        {
-            
+        if ($form->isSubmitted() && $form->isValid()) {
+
             $em->persist($pin);
             $em->flush();
 
             return $this->redirectToRoute('app_pins_show', ['id' => $pin->getId()]);
         }
 
-       return $this->render('pins/create.html.twig', [
-           'form' => $form->createView()
-       ]);
+        return $this->render('pins/create.html.twig', [
+            'form' => $form->createView()
+        ]);
     }
 
-      /** 
-    *@Route("/pins/{id<[0-9]+>}/edit", name="app_pins_edit", methods={"GET", "PUT"})
-    */
+    /** 
+     *@Route("/pins/{id<[0-9]+>}/edit", name="app_pins_edit", methods={"GET", "PUT"})
+     */
     public function edit(Request $request, EntityManagerInterface $em, Pin $pin): Response
     {
         $form = $this->createForm(PinType::class, $pin, [
-            'method' =>'PUT'
+            'method' => 'PUT'
         ]);
 
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid())
-        {
-            
+        if ($form->isSubmitted() && $form->isValid()) {
+
             $em->flush();
             return $this->redirectToRoute('app_home');
         }
@@ -73,26 +71,36 @@ class PinsController extends AbstractController
         ]);
     }
 
+
+    /* public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
+        {
+            if ($this->isCsrfTokenValid('pin_deletion' . $pin->getId(), $request->request->get('csrf_token'))) {
+                $this->em->remove($pin);
+                $this->em->flush();
+            }
+            return $this->redirectToRoute('app_home');
+        } */
+
     /**
      * @param Pin $pin
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
-        public function delete(Pin $pin, EntityManagerInterface $em): Response
-        {
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
+    {
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token'))) {
             $em->remove($pin);
             $em->flush();
-            return $this->redirectToRoute('app_home');
         }
+        return $this->redirectToRoute('app_home');
+    }
 
     /**
      * @Route("/pins/{id<[0-9]+>}", name="app_pins_show")
      */
     public function show(Pin $pin): Response
     {
-     
+
         return $this->render('pins/show.html.twig', compact('pin'));
     }
-
-
 }
